Client_type is now reported as a string.
client_Type is passed to the analytics module as a number. It needs to
be converted to a string to be a valid BigQuery entry.

Adds some additional checking in the Analytics unittest code to validate
BigQuery types in the tests. This should catch this type of bug in the future.

Fixes #139